General code cleanup

+++ src/chatcontrol/Storage.java
@@ -5,8 +5,8 @@ import org.bukkit.Location;
 public class Storage {
 	public String lastMessage = "";
 	public String lastCommand = "";
-	public Long lastMessageTime = (long) 0;
-	public Long lastCommandTime = (long) 0;
+	public long lastMessageTime = 0;
+	public long lastCommandTime = 0;
 	public Location loginLocation = null;
 	public String lastSignText = "";
 }
